refactor(boot): move semantic status types to core
Move BootPhaseStatus, BootRunStatus, BootPhaseSnapshot and BootRunSnapshot
from spring-application.bootstrap to core.enums/dto.boot, where they belong
as framework-agnostic domain vocabulary.

Rename Phase2Mode -> BootFailureMode and Phase2AccountQueryPolicy ->
BootAccountQueryPolicy to reflect boot-scoped intent rather than phase
position, keeping values (WARN_ONLY/FAIL_FAST, SKIP/FAIL) unchanged.

Spring consumers (BootStatusTracker, BootMetricsRecorder, BootOrchestrator,
RunnerBootPhase2Properties, BootController) and the three BootOrchestrator
test classes are updated to reference the new core locations.
