+++ core/src/main/java/com/marmitt/core/dto/boot/BootPhaseSnapshot.java
@@ -1,4 +1,6 @@
-package com.marmitt.application.spring.bootstrap;
+package com.marmitt.core.dto.boot;
+
+import com.marmitt.core.enums.BootPhaseStatus;
 
 import java.time.Instant;
