+++ core/src/main/java/com/marmitt/core/enums/BootAccountQueryPolicy.java
@@ -1,0 +1,6 @@
+package com.marmitt.core.enums;
+
+public enum BootAccountQueryPolicy {
+    SKIP,
+    FAIL
+}
