+++ core/src/main/java/com/marmitt/core/enums/BootFailureMode.java
@@ -1,0 +1,6 @@
+package com.marmitt.core.enums;
+
+public enum BootFailureMode {
+    WARN_ONLY,
+    FAIL_FAST
+}
