--- spring-application/src/main/java/com/marmitt/application/spring/bootstrap/Phase2AccountQueryPolicy.java
@@ -1,6 +1,0 @@
-package com.marmitt.application.spring.bootstrap;
-
-public enum Phase2AccountQueryPolicy {
-    SKIP,
-    FAIL
-}
