--- spring-application/src/main/java/com/marmitt/application/spring/bootstrap/Phase2Mode.java
@@ -1,6 +1,0 @@
-package com.marmitt.application.spring.bootstrap;
-
-public enum Phase2Mode {
-    WARN_ONLY,
-    FAIL_FAST
-}
