+++ spring-application/src/test/java/com/marmitt/application/spring/bootstrap/BootOrchestratorObservabilityTest.java
@@ -11,6 +11,9 @@ import com.marmitt.core.dto.portfolio.PortfolioZombieDetectionResult;
 import com.marmitt.core.enums.AccountingPolicyType;
 import com.marmitt.core.enums.DlqReason;
 import com.marmitt.core.enums.ExecutionPolicy;
+import com.marmitt.core.enums.BootFailureMode;
+import com.marmitt.core.enums.BootPhaseStatus;
+import com.marmitt.core.enums.BootRunStatus;
 import com.marmitt.core.enums.RunnerStatus;
 import com.marmitt.core.ports.outbound.repository.DeadLetterEntryRepositoryPort;
 import com.marmitt.core.ports.outbound.repository.ExchangeAdapterRepositoryPort;
@@ -48,7 +51,7 @@ class BootOrchestratorObservabilityTest {
         SimpleMeterRegistry meterRegistry = new SimpleMeterRegistry();
         CapturingEventPublisher eventPublisher = new CapturingEventPublisher();
         BootOrchestrator orchestrator = newOrchestrator(
-                Phase2Mode.WARN_ONLY,
+                BootFailureMode.WARN_ONLY,
                 portfolio,
                 runner,
                 detected,
@@ -76,14 +79,14 @@ class BootOrchestratorObservabilityTest {
                 .tag("phase", "phase2.zombie")
                 .tag("status", "DETECTED")
                 .tag("exchange", "MOCK")
-                .tag("mode", Phase2Mode.WARN_ONLY.name())
+                .tag("mode", BootFailureMode.WARN_ONLY.name())
                 .counter()
                 .count());
         assertEquals(1L, meterRegistry.get("boot.phase.portfolio.duration")
                 .tag("phase", "phase2.zombie")
                 .tag("status", "DETECTED")
                 .tag("exchange", "MOCK")
-                .tag("mode", Phase2Mode.WARN_ONLY.name())
+                .tag("mode", BootFailureMode.WARN_ONLY.name())
                 .timer()
                 .count());
         assertEquals(0, eventPublisher.events().size());
@@ -108,7 +111,7 @@ class BootOrchestratorObservabilityTest {
         SimpleMeterRegistry meterRegistry = new SimpleMeterRegistry();
         CapturingEventPublisher eventPublisher = new CapturingEventPublisher();
         BootOrchestrator orchestrator = newOrchestrator(
-                Phase2Mode.FAIL_FAST,
+                BootFailureMode.FAIL_FAST,
                 portfolio,
                 runner,
                 detected,
@@ -136,7 +139,7 @@ class BootOrchestratorObservabilityTest {
                 .tag("phase", "phase2.zombie")
                 .tag("status", "DETECTED")
                 .tag("exchange", "MOCK")
-                .tag("mode", Phase2Mode.FAIL_FAST.name())
+                .tag("mode", BootFailureMode.FAIL_FAST.name())
                 .counter()
                 .count());
         assertEquals(1.0d, meterRegistry.get("boot.failfast.total")
@@ -154,7 +157,7 @@ class BootOrchestratorObservabilityTest {
         assertNotNull(event.timestamp());
     }
 
-    private static BootOrchestrator newOrchestrator(Phase2Mode mode,
+    private static BootOrchestrator newOrchestrator(BootFailureMode mode,
                                                     Portfolio portfolio,
                                                     StrategyRunner runner,
                                                     PortfolioZombieDetectionResult zombieResult,
